Fixed minor bugs & cleanup
