+++ .gitignore
@@ -5,6 +5,7 @@
 /build
 /captures
 /desktop-app/build
+/desktop-app/out
 .externalNativeBuild
 .cxx
 local.properties
