+++ desktop-app/src/main/java/com/arsvechkarev/vaultdesktop/EntriesHolder.kt
@@ -1,0 +1,7 @@
+package com.arsvechkarev.vaultdesktop
+
+import com.arsvechkarev.vaultdesktop.model.Entries
+
+object EntriesHolder {
+  var entries: Entries? = null
+}
