--- desktop-app/src/main/kotlin/com/arsvechkarev/vaultdesktop/EntriesHolder.kt
@@ -1,7 +1,0 @@
-package com.arsvechkarev.vaultdesktop
-
-import com.arsvechkarev.vaultdesktop.model.Entries
-
-object EntriesHolder {
-  var entries: Entries? = null
-}
