--- desktop-app/src/main/kotlin/com/arsvechkarev/vaultdesktop/model/Entries.kt
@@ -1,12 +1,0 @@
-package com.arsvechkarev.vaultdesktop.model
-
-import com.google.gson.annotations.SerializedName
-
-data class Entries(
-  @SerializedName("passwords")
-  val passwords: List<Password>,
-  @SerializedName("creditCards")
-  val creditCards: List<CreditCard>,
-  @SerializedName("plainTexts")
-  val plainTexts: List<PlainText>
-)
