--- desktop-app/src/main/kotlin/com/arsvechkarev/vaultdesktop/style/Colors.kt
@@ -1,7 +1,0 @@
-package com.arsvechkarev.vaultdesktop.style
-
-object Colors {
-  
-  const val ACCENT = "#3C80AE"
-  const val ERROR = "#E90202"
-}
